after_mech/requirements.txt
asgiref==3.9.1
channels==4.3.1
channels_redis==4.3.0
Django==4.2.23
djangorestframework==3.16.1
msgpack==1.1.1
pillow==11.3.0
redis==6.4.0
sqlparse==0.5.3
daphne>=4

python-dotenv>=1.0.0
django-axes>=6.0.0
django-cors-headers>=4.3.0
httpx>=0.25.0
django-ratelimit>=4.0.0
langdetect>=1.0.9
Unidecode>=1.3.8
google-cloud-vision>=3.7.0
django-redis>=5.3.0
whitenoise>=6.6.0
